EJERCICIO 6 REALIZADO Y RESUELTO

+++ Ejercicio04/ejercisio04/src/main/java/com/mycompany/ejercisio04/Ejercisio04.java
@@ -6,7 +6,7 @@ public class Ejercisio04 {
         int var2 = 1;
         int var3 = 3;
         int var4 = 2;
-        double respuesta = (double)var1/(var2 + var3) + var4;
+        double respuesta = var1/(var2 + var3) + var4;
              System.out.println("LA RESPUESTA ES:" + respuesta);
         //ESTO ESTA REALISADO x2
     }
